Check games folder of computer and choose one
This is the first time I have made something by having and idea and figuring out how to do it without going with a tutorial. I am sure there is a hundred much better ways to do something like this and I am happy to experiment and try to find how. Right now I am a caveman and this is my rock hammer.
I know currently it presents some non game like steam controller configs so this is version 0 and I'm hoping to make a more refined one over time.