Source FIle
This is completely monkey stuff. I can already think of maybe five ways to do this way better. Regardless, I am fairly excited as it's the first time I had an idea and made something without a tutorial or anything just knowing what I need to do and making it researching what I needed to do to make it happen.